+++ package.json
@@ -39,10 +39,12 @@
   },
   "dependencies": {
     "ejs": "3.1.10",
-    "snarkjs": "0.7.5"
+    "snarkjs": "0.7.5",
+    "@iden3/binfileutils": "0.0.12",
+    "ffjavascript": "0.3.1",
+    "readline": "1.3.0"
   },
   "devDependencies": {
-    "@iden3/binfileutils": "^0.0.12",
     "@nomicfoundation/hardhat-ethers": "3.0.5",
     "@nomiclabs/hardhat-vyper": "^3.0.7",
     "@solarity/solidity-lib": "3.0.0-rc.3",
@@ -54,13 +56,11 @@
     "chai": "^4.4.1",
     "chai-as-promised": "^7.1.1",
     "ethers": "^6.11.1",
-    "ffjavascript": "^0.3.1",
     "hardhat": "2.22.7",
     "husky": "^9.0.11",
     "mocha": "^10.3.0",
     "nyc": "^15.1.0",
     "prettier": "^3.2.5",
-    "readline": "^1.3.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5"
   }
